test: add microk8s cloud in conftest.py
