+++ tox.ini
@@ -17,8 +17,6 @@ all_path = {[vars]src_path} {[vars]tests_path}
 set_env =
     PYTHONPATH = {[vars]src_path}:{tox_root}/lib
     PY_COLORS = 1
-    lxd: CLOUD=localhost
-    k8s: CLOUD=microk8s
 allowlist_externals =
     poetry
